Enable usage of additional_metadata inside meta (#1)
### Description

Render `meta.aditional_meta` in dbt docs. `meta_additional` should be a
list of lists, where each inner list contains dictionaries with one
key-val pair. The index of the outer list defines the position of the
row, the index of the inner list(s) defines the position inside the row.

```
    meta:
      meta_stuff: meta_stuff
      additional_docs:
        - - foo: bar
          - baz: qux
        - - fiz: buz
          - bla: bla
```

---------

Co-authored-by: Matthieu Caneill <matthieucan@users.noreply.github.com>